Update README.md
@mmtrt Thought to put it here to avoid people creating issues and make it easier to use the snap. I'm very new to markdown so please edit the structure if it doesn't suit you.

p.s. There was a link to your Patreon here, and I believe it should be here. Before I've managed to set up a Patreon account it was removed. Please bring it back, or give me your Patreon username and I will put it in here using in-build GitHub funding option.